pyproject.toml
[project]
name = "online-learning-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Tem Temich (Deckop)",email = "khxkxmxv@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "djangorestframework-stubs (>=3.16.8,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "stripe (>=14.4.1,<15.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
